users/urls.py
from django.urls import path
from rest_framework.permissions import AllowAny

from users.apps import UsersConfig
from users.views import UserRetrieveUpdateAPIView, UserListAPIView, UserCreateAPIView
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView

app_name = UsersConfig.name

urlpatterns = [
    path("retrieve_update/<int:pk>/", UserRetrieveUpdateAPIView.as_view(), name="users_retrieve_update"),
    path('register/', UserCreateAPIView.as_view(), name='register'),
    path('login/', TokenObtainPairView.as_view(permission_classes=(AllowAny,)), name='token_obtain_pair'),
    path('token/refresh/', TokenRefreshView.as_view(permission_classes=(AllowAny,)), name='token_refresh'),

   
    path("list/", UserListAPIView.as_view(), name="users_list"),
]