pyproject.toml
[tool.poetry]
name = "kyrs-django-rest"
version = "0.1.0"
description = ""
authors = ["nikolyai-tarasov <kolya.tarasov333@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.5"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
ipython = "^8.31.0"
psycopg2-binary = "^2.9.10"
django-filter = "^24.3"
drf-yasg = "^1.21.8"
djangorestframework-simplejwt = "^5.4.0"
celery = "^5.4.0"
django-celery-beat = "^2.7.0"
requests = "^2.32.3"
pillow = "^11.1.0"
coverage = "^7.6.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
